added clustering, UMAP, and plotting fxns
